check-bugs: more correct tag matching
When a commit is in multiple tags (should be rare), the last tag was
used for displaying. This is normally incorrect, as the commit was
merged in a not most recent release.

In order to pick the oldest release, the tags are split in 'testing' and
'release' lists. Only when the commit is not in the 'release' list, the
list of 'testing' tags is consulted to pick a tag.

Reported-by: Shyamsundar Ranganathan <srangana@redhat.com>
Signed-off-by: Niels de Vos <ndevos@redhat.com>